requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-elasticsearch-dsl==7.2.2
django-import-export==2.8.0
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
urllib3==1.22
xlrd==2.0.1
xlwt==1.3.0
